requirements.txt
# Google ADK and core dependencies
google-adk>=0.1.0
google-cloud-aiplatform>=1.70.0
google-generativeai>=0.8.0

# Web framework and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data handling
pydantic>=2.5.0
requests>=2.31.0

# Geolocation services
geopy>=2.4.0

# Environment and configuration
python-dotenv>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

